Refactor(logging): Migrate logger configuration and factory to new domain structure
- Deleted legacy logger configuration module and associated factory class to streamline the codebase.
- Created new logging domain structure under app/logging/ with config.py and factory.py.
- Updated main settings to remove dependency on the old logger configuration.
- Adjusted imports in various modules to reference the new logging domain.
- Implemented @lru_cache for singleton pattern in the new logging config.
- Enhanced maintainability and reduced complexity in logging management.
- Verified functionality of all components post-refactor, ensuring logging works correctly across the application.
