+++ app/pipeline/event/delete/executor.py
@@ -12,7 +12,7 @@ from app.calendar.service import GoogleCalendarService
 from app.core.exceptions import CalServiceError, ErrorMessages, ValidationError
 from app.core.node import Node
 from app.core.schema.task import TaskContext
-from app.services.logger_factory import logger
+from app.logging.factory import logger
 
 
 class DeleteEventExecutor(Node):
