+++ app/worker/tasks.py
@@ -4,8 +4,8 @@ from app.api.schema import EventSchema
 from app.database.event import Event
 from app.database.repository import GenericRepository
 from app.database.session import get_db_session
+from app.logging.factory import logger, set_request_id
 from app.pipeline.pipeline import CalendarPipeline
-from app.services.logger_factory import logger, set_request_id
 from app.worker.celery_app import celery_app
 
 """
